[Silver V] Title: 공포의 면담실, Time: 568 ms, Memory: 81016 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/30088. 공포의 면담실/공포의 면담실.java
@@ -8,7 +8,7 @@ public class Main {
         StringTokenizer st;
 
         int N = Integer.parseInt(br.readLine());
-        int[] arr = new int[N];
+        long[] arr = new long[N];
 
         for(int i = 0; i < N; i++){
             st = new StringTokenizer(br.readLine());
@@ -32,7 +32,7 @@ public class Main {
             }
         }
 
-        int res = 0;
+        long res = 0;
         for(int i = 0; i < N; i++){
            res += arr[i];
         }
